requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
pandas
numpy
scikit-learn
python-dotenv
requests